Add support to print help for commands.
Add a .helpmsg field that contains a string with information of how
the commmand should be used. The help callback that prints the message
is executed with the 'h' character following the command; i.e.

pwmfreq h

The user is responsible for providing the print function and must be
named tinycmd_printf.

If help message printing is not needed, then it can be compiled out
using the TINYCMD_HELP_ENABLE macro.